fix(RoomescapeRepositoryImpl): 도메인 추가에 따른 CRUD 쿼리 변경
